Crate gol

Experimental Rust library for implementing Conway's Game of Life

See also gol-tcod for example code consuming this library.

Modules

grid
The grid module provides Grid for working with a "torus world" grid of cells.

plaintext
Module for parsing the Plaintext Game of Life file format into a Grid.

rules
Module containing functions implementing Game of Life rulesets.

world
Structures and functions used for constructing a Game of Live World which combines a Grid and a rule set.