Struct gol::grid::Grid

pub struct Grid {
    // some fields omitted
}

An addressable grid of Cells

Provides a number of functions for constructing, modifying and walking Cell grids.

Methods

impl Grid

fn from_raw(width: usize, height: usize, state: Vec<Cell>) -> Grid

Constructs a Grid from raw components

fn from_fn<F>(width: usize, height: usize, f: F) -> Grid where F: FnMut(usize, usize) -> Cell

Constructs a Grid of width and height using a factory function.

fn create_dead(width: usize, height: usize) -> Grid

Constructs a dead grid of width and height

fn width(&self) -> usize

Gets the width of this Grid

fn height(&self) -> usize

Gets the height of this Grid

fn cell_at(&self, x: usize, y: usize) -> &Cell

Returns a reference to the Cell at the given coordinates

fn set_cell(&mut self, x: usize, y: usize, cell: Cell)

Overwrites the Cell at the given coordinates with the given value

fn write_cells(&mut self, x: usize, y: usize, data: &Grid)

Overwrite the cells starting at coords (x, y) with the data in the given Grid If any coordinates are outside the grid no action is taken.

fn iter_rows(&self) -> RowIter

Returns an iterator over rows in this Grid

fn iter_cells(&self) -> CellIter

Returns an iterator over Cells in this Grid