Module gol::plaintext

Module for parsing the Plaintext Game of Life file format into a Grid.

Structs

Padding
Describes padding in the order top, right, bottom, left

PlainText
Struct for the contents of a Plaintext format Game of Life file.

Enums

Error
Represents any errors which occur during the Plaintext parsing process

Functions

parse_plaintext
Parses the Plaintext format from a buffered stream

Type Definitions

ParseResult
Represents the result of a Plaintext parse operation