Struct gol::plaintext::PlainText

pub struct PlainText {
    pub name: String,
    pub comment: String,
    pub data: Grid,
}

Struct for the contents of a Plaintext format Game of Life file.

Optional padding syntax

The plaintext parser also provides a Padding extension, e.g.:

!Name: Example
!Padding: 5,10,5
!
.O.
O.O
.O.

Resulting in the following padding being applied to the result:

Top 	Right 	Bottom 	Left

5 	10 	5 	10

Fields

name

comment

data