gol::rules::RulesFn

type RulesFn = fn(cell: &Cell, neighbours: usize) -> Cell;

Describes a static ruleset function.

This function accepts a current cell state and the count of neighbours that cell has and returns a new cell state.