Struct gol::Grid

pub struct Grid {
    // some fields omitted
}

An addressable grid of Cells

Provides a number of functions for constructing, modifying and walking Cell grids.

Methods

impl Grid

fn from_raw(width: usize, height: usize, state: Vec<Cell>) -> Grid

Constructs a Grid from raw components

fn from_fn<F>(width: usize, height: usize, f: F) -> Grid where F: FnMut((usize, usize)) -> Cell

Constructs a Grid of width and height using a factory function.

fn create_dead(width: usize, height: usize) -> Grid

Constructs a dead grid of width and height

fn create_random<R: Rng>(rng: &mut R, width: usize, height: usize) -> Grid

Constructs a random grid of width and height

fn iter_rows(&self) -> RowIterator

Returns an iterator over rows in this Grid

fn cell_at(&self, x: usize, y: usize) -> &Cell

Returns a reference to the Cell at the given coordinates

fn set_cell(&mut self, x: usize, y: usize, cell: Cell)

Overwrites the Cell at the given coordinates with the given value