Struct gol::World

pub struct World {
    // some fields omitted
}

Represents a Game of Life Grid + generation

Methods

impl World

fn new(grid: Grid) -> World

Constructs a new World with the given Grid

fn generation(&self) -> i64

Gets the current generation for this World

fn width(&self) -> usize

Gets the width of this World

fn height(&self) -> usize

Gets the height of this World

fn step_mut(&mut self)

Executes a single step of this World in place

fn step(&self) -> World

Executes a single step of this World, and returns a new world

fn write_cells(&mut self, x: usize, y: usize, data: &Grid)

Overwrite the cells starting at coords (x, y) with the data in the given Grid

fn iter_rows(&self) -> RowIterator

Returns an iterator over rows in this world