Struct gol::world::World

pub struct World {
    // some fields omitted
}

Provides hosting for a basic Game of Life simulation. Includes functions for modifying the world and stepping the simulation both immutably and in-place.

Methods

impl World

fn new(grid: Grid) -> World

Constructs a new World with the given Grid

fn new_with_rules(grid: Grid, rules: RulesFn) -> World

Constructs a new World with the given Grid and RulesFn

fn generation(&self) -> i64

Gets the current generation for this World

fn width(&self) -> usize

Gets the width of this World

fn height(&self) -> usize

Gets the height of this World

fn step_mut(&mut self)

Executes a single step of this World in place

fn step(&self) -> World

Executes a single step of this World and returns a new, modified world

fn write_cells(&mut self, x: usize, y: usize, data: &Grid)

Overwrite the cells starting at coords (x, y) with the data in the given Grid

fn iter_rows(&self) -> RowIter

Returns an iterator over rows in this world